+++ src/main/app/src/app/core/service/util.service.ts
@@ -6,27 +6,10 @@
 import { BreakpointObserver, Breakpoints } from "@angular/cdk/layout";
 import { ComponentType } from "@angular/cdk/portal";
 import { DOCUMENT } from "@angular/common";
-import {
-  Component,
-  Inject,
-  Injectable,
-  Optional,
-  Signal,
-  computed,
-  inject,
-} from "@angular/core";
+import { Inject, Injectable, Optional, Signal } from "@angular/core";
 import { MatDialog } from "@angular/material/dialog";
-import {
-  MatProgressBar,
-  ProgressBarMode,
-} from "@angular/material/progress-bar";
-import {
-  MAT_SNACK_BAR_DATA,
-  MatSnackBar,
-  MatSnackBarConfig,
-  MatSnackBarLabel,
-  MatSnackBarRef,
-} from "@angular/material/snack-bar";
+import { MatSnackBar, MatSnackBarConfig } from "@angular/material/snack-bar";
+import { ProgressSnackbar } from "src/app/shared/progress-snackbar/progress-snackbar.component";
 import { Title } from "@angular/platform-browser";
 import { Router } from "@angular/router";
 import { TranslateService } from "@ngx-translate/core";
@@ -260,38 +243,4 @@ export class UtilService {
     link.remove();
     window.URL.revokeObjectURL(link.href);
   }
-}
-
-@Component({
-  standalone: true,
-  imports: [MatSnackBarLabel, MatProgressBar],
-  template: `
-    <div matSnackBarLabel>{{ data.message }}</div>
-    <mat-progress-bar
-      [mode]="progressMode()"
-      [value]="data.progressPercentage()"
-    ></mat-progress-bar>
-  `,
-  styles: `
-    .mat-mdc-progress-bar {
-      --mdc-linear-progress-active-indicator-color: var(
-        --mat-snack-bar-button-color
-      );
-      --mdc-linear-progress-track-color: rgba(255, 64, 129, 0.25);
-      position: absolute;
-      bottom: 0;
-    }
-  `,
-})
-class ProgressSnackbar {
-  snackBarRef = inject(MatSnackBarRef);
-  progressMode = computed<ProgressBarMode>(() => {
-    const percentage = this.data.progressPercentage();
-    return percentage === 100 || percentage === 0 ? "query" : "determinate";
-  });
-
-  constructor(
-    @Inject(MAT_SNACK_BAR_DATA)
-    public data: { progressPercentage: Signal<number>; message: string },
-  ) {}
 }
