+++ src/main/app/src/app/shared/progress-snackbar/progress-snackbar.component.less
@@ -1,0 +1,13 @@
+/*
+ * SPDX-FileCopyrightText: 2024 Lifely
+ * SPDX-License-Identifier: EUPL-1.2+
+ */
+
+.mat-mdc-progress-bar {
+  --mdc-linear-progress-active-indicator-color: var(
+    --mat-snack-bar-button-color
+  );
+  --mdc-linear-progress-track-color: rgba(255, 64, 129, 0.25);
+  position: absolute;
+  bottom: 0;
+}
